Rename auth to identity and shuffle folder hierarchy to match DDD
