+++ ui/src/components/identity/login/__tests__/login.test.js
@@ -1,8 +1,8 @@
 import Login from '../';
 import React from 'react';
 import Sinon from 'sinon';
-import {InlineErrorNotification} from 'components/inline_notifications';
-import {PrimaryButton} from 'components/buttons';
+import {InlineErrorNotification} from 'components/shared/inline_notifications';
+import {PrimaryButton} from 'components/shared/buttons';
 
 const sandbox = Sinon.createSandbox();
