+++ ui/src/components/identity/login/index.js
@@ -3,8 +3,8 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import withField from 'hocs/with_field';
 import {Field} from 'redux-form';
-import {PrimaryButton} from 'components/buttons';
-import {InlineErrorNotification} from 'components/inline_notifications';
+import {PrimaryButton} from 'components/shared/buttons';
+import {InlineErrorNotification} from 'components/shared/inline_notifications';
 
 const InputField = withField('input');
