+++ ui/src/components/shared/inline_notifications/inline_notifications.scss
