+++ ui/src/constants/actions.js
@@ -1,8 +1,8 @@
 export default {
   Api: {CALL: 'api/CALL'},
-  Auth: {
-    LOGIN_REQUEST: 'auth/LOGIN_REQUEST',
-    LOGIN_FAILURE: 'auth/LOGIN_FAILURE',
-    LOGIN_SUCCESS: 'auth/LOGIN_SUCCESS'
+  Identity: {
+    LOGIN_REQUEST: 'identity/LOGIN_REQUEST',
+    LOGIN_FAILURE: 'identity/LOGIN_FAILURE',
+    LOGIN_SUCCESS: 'identity/LOGIN_SUCCESS'
   }
 };
