+++ ui/src/reducers/identity.js
@@ -15,12 +15,12 @@ const initialState = {isAuthenticated: false};
 const {TOKEN_KEY} = LocalStorageConstants;
 
 export default createReducer(initialState, {
-  [ActionTypes.Auth.LOGIN_FAILURE]: state => {
+  [ActionTypes.Identity.LOGIN_FAILURE]: state => {
     localStorage.removeItem(TOKEN_KEY);
     return Object.assign({}, state, {isAuthenticated: false});
   },
 
-  [ActionTypes.Auth.LOGIN_SUCCESS]: (state, {payload}) => {
+  [ActionTypes.Identity.LOGIN_SUCCESS]: (state, {payload}) => {
     // TODO: Ugh. Let's just cookie / session this mother fucker.
     localStorage.setItem(TOKEN_KEY, payload.token);
     return Object.assign({}, state, {isAuthenticated: true});
