+++ ui/src/reducers/index.js
@@ -1,4 +1,4 @@
-import auth from './auth';
+import identity from './identity';
 import {reducer as form} from 'redux-form';
 
-export default {auth, form};
+export default {identity, form};
